refactor: split zero-sidebar.tsx into focused sub-components
Extract AccountDropdown, RecentChatSection, TalkToSection,
OverlayScrollArea, and SidebarUpgradeCard into dedicated files,
reducing the main file from 1,678 to 607 lines.

Closes #7834
