+++ turbo/apps/platform/src/views/zero-page/zero-sidebar.tsx
@@ -13,48 +13,20 @@ import {
   IconChartLine,
   IconLayoutGrid,
   IconCalendar,
-  IconAdjustmentsHorizontal,
-  IconUser,
   IconUsers,
-  IconLogout,
-  IconPlus,
-  IconChevronRight,
-  IconSwitchHorizontal,
-  IconSearch,
-  IconX,
   IconEdit,
+  IconChevronRight,
   IconLayoutSidebarLeftCollapse,
-  IconDatabaseExport,
   IconPlug,
-  IconTrash,
 } from "@tabler/icons-react";
-import { FeatureSwitchKey, type ChatThreadListItem } from "@vm0/core";
+import { FeatureSwitchKey } from "@vm0/core";
 import {
-  DropdownMenu,
-  DropdownMenuTrigger,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuSeparator,
-  DropdownMenuSub,
-  DropdownMenuSubTrigger,
-  DropdownMenuSubContent,
   Tooltip,
   TooltipContent,
   TooltipProvider,
   TooltipTrigger,
-  Button,
 } from "@vm0/ui";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-  DialogFooter,
-} from "@vm0/ui/components/ui/dialog";
-import { Skeleton } from "@vm0/ui/components/ui/skeleton";
 import slackIcon from "./components/settings/icons/slack.svg";
-import { clerk$, user$, resolveWebOrigin } from "../../signals/auth.ts";
 import { detach, Reason } from "../../signals/utils.ts";
 import {
   activeRoute$,
@@ -67,9 +39,7 @@ import {
   handleZeroAccountAction$,
   navigateToChat$,
   sidebarChatAgentId$,
-  isChatRoute,
   type SidebarNavId,
-  type ZeroAccountAction,
 } from "../../signals/zero-page/zero-nav.ts";
 import type { RouteKey } from "../../signals/route-paths.ts";
 import {
@@ -77,36 +47,18 @@ import {
   defaultAgentId$,
 } from "../../signals/zero-page/zero-agent-name.ts";
 import { zeroSubagents$ } from "../../signals/zero-page/zero-agents.ts";
-import { reloadAgents$ } from "../../signals/zero-page/agents-list.ts";
 import {
   chatThreads$,
   createNewChatThread$,
   creatingNewSession$,
-  deleteChatThread$,
 } from "../../signals/zero-page/zero-chat.ts";
 import {
   pinnedAgentIds$,
   updatePinnedAgentIds$,
 } from "../../signals/zero-page/zero-pinned-agents.ts";
 import {
-  sidebarSearchOpen$,
-  sidebarSearchTerm$,
-  setSidebarSearchOpen$,
-  setSidebarSearchTerm$,
   managePinnedDialogOpen$,
   setManagePinnedDialogOpen$,
-  pendingDeleteThreadId$,
-  setPendingDeleteThreadId$,
-  sessionListCollapsed$,
-  setSessionListCollapsed$,
-  chatListOpen$,
-  setChatListOpen$,
-  agentCardCollapsed$,
-  setAgentCardCollapsed$,
-  thumbStyle$,
-  setThumbStyle$,
-  hovering$,
-  setHovering$,
   isScrolled$,
   setIsScrolled$,
   manageSectionCollapsed$,
@@ -114,32 +66,22 @@ import {
 } from "../../signals/zero-page/zero-sidebar-state.ts";
 import { VM0ClerkProvider } from "../clerk/clerk-provider.tsx";
 import { ZeroOrgSwitcher } from "./zero-org-switcher.tsx";
-import {
-  AgentAvatarImg,
-  useAgentAvatar,
-  type SubagentInfo,
-} from "./zero-sidebar-shared.tsx";
+import { useAgentAvatar, type SubagentInfo } from "./zero-sidebar-shared.tsx";
 import { Link } from "../router/link.tsx";
 import { featureSwitch$ } from "../../signals/external/feature-switch.ts";
-import { apiBaseForNavigation$ } from "../../signals/fetch.ts";
-import { billingStatusAsync$ } from "../../signals/zero-page/billing.ts";
-import planProImg from "./components/org-manage/assets/plan-pro.webp";
-import planTeamImg from "./components/org-manage/assets/plan-team.webp";
-import {
-  setActiveTab$,
-  setBillingSubPage$,
-} from "../../signals/zero-page/settings/org-manage-tabs-state.ts";
-import { setOrgManageDialogOpen$ } from "../../signals/zero-page/settings/org-manage-dialog.ts";
-import { isOrgAdmin$ } from "../../signals/org.ts";
 import { slackOrgScopeMismatch$ } from "../../signals/zero-page/zero-slack.ts";
 import { BillingDialog } from "./billing-dialog.tsx";
-import {
-  ChatListDialog,
-  ManagePinnedAgentsDialog,
-} from "./zero-sidebar-dialogs.tsx";
+import { ManagePinnedAgentsDialog } from "./zero-sidebar-dialogs.tsx";
+
+import { AccountDropdown } from "./zero-sidebar-account.tsx";
+import { RecentChatSection } from "./zero-sidebar-chat-list.tsx";
+import { TalkToSection } from "./zero-sidebar-pinned.tsx";
+import { OverlayScrollArea } from "./zero-sidebar-scroll.tsx";
+import { SidebarUpgradeCard } from "./zero-sidebar-upgrade.tsx";
 
 // Re-export shared types/components for backward compatibility
 export { useAgentAvatar } from "./zero-sidebar-shared.tsx";
+export { AccountDropdown } from "./zero-sidebar-account.tsx";
 
 type NavIcon = (props: { size?: number; className?: string }) => ReactNode;
 
@@ -212,1024 +154,6 @@ const FOOTER_NAV = [
     featureGate: FeatureSwitchKey.Usage,
   },
 ] as const satisfies readonly FooterNavItem[];
-
-interface SessionAccount {
-  sessionId: string;
-  name: string;
-  email: string;
-  initial: string;
-  imageUrl: string | undefined;
-  isActive: boolean;
-}
-
-function AccountAvatar({
-  imageUrl,
-  name,
-  initial,
-  size = "sm",
-}: {
-  imageUrl: string | undefined;
-  name: string;
-  initial: string;
-  size?: "sm" | "md";
-}) {
-  const dim = size === "md" ? "h-9 w-9" : "h-8 w-8";
-  const textSize = size === "md" ? "text-sm" : "text-xs";
-  if (imageUrl) {
-    return (
-      <div className={`${dim} shrink-0 rounded-xl overflow-hidden`}>
-        <img src={imageUrl} alt={name} className="h-full w-full object-cover" />
-      </div>
-    );
-  }
-  return (
-    <div
-      className={`${dim} rounded-xl bg-orange-200/95 dark:bg-orange-300/80 flex items-center justify-center text-orange-900 dark:text-orange-950 ${textSize} font-medium shrink-0`}
-    >
-      {initial}
-    </div>
-  );
-}
-
-function useAccountSessions() {
-  const clerkLoadable = useLoadable(clerk$);
-  const userLoadable = useLoadable(user$);
-  const user = userLoadable.state === "hasData" ? userLoadable.data : null;
-  const clerk = clerkLoadable.state === "hasData" ? clerkLoadable.data : null;
-
-  const currentSessionId = clerk?.session?.id;
-  const accounts: SessionAccount[] = (clerk?.client?.sessions ?? [])
-    .filter((s) => {
-      return s.status === "active";
-    })
-    .map((s) => {
-      return {
-        sessionId: s.id,
-        name: s.user?.fullName ?? "User",
-        email: s.user?.primaryEmailAddress?.emailAddress ?? "",
-        initial: s.user?.fullName
-          ? s.user.fullName.charAt(0).toUpperCase()
-          : "U",
-        imageUrl: s.user?.imageUrl,
-        isActive: s.id === currentSessionId,
-      };
-    });
-
-  return { user, clerk, accounts };
-}
-
-export function AccountDropdown({
-  onAccountAction,
-  collapsed = false,
-  hidePreferences = false,
-}: {
-  onAccountAction?: (action: ZeroAccountAction) => void;
-  collapsed?: boolean;
-  hidePreferences?: boolean;
-}) {
-  const { user, clerk, accounts } = useAccountSessions();
-  const features = useLastResolved(featureSwitch$);
-  const apiBase = useGet(apiBaseForNavigation$);
-  const showExportData = features?.[FeatureSwitchKey.DataExport] ?? false;
-  const accountName = user?.fullName ?? "User";
-  const accountEmail = user?.primaryEmailAddress?.emailAddress ?? "";
-  const accountInitial = accountName.charAt(0).toUpperCase();
-
-  const current = accounts.find((a) => {
-    return a.isActive;
-  });
-  const others = accounts.filter((a) => {
-    return !a.isActive;
-  });
-  const hasOthers = others.length > 0;
-
-  const handleAccountAction = (action: ZeroAccountAction) => {
-    if (action === "signout") {
-      const sessionId = clerk?.session?.id;
-      const signInUrl = `${resolveWebOrigin()}/sign-in?redirect_url=${encodeURIComponent(location.href)}`;
-      detach(
-        clerk?.signOut({ sessionId, redirectUrl: signInUrl }),
-        Reason.DomCallback,
-      );
-      return;
-    }
-    if (action === "manage") {
-      detach(clerk?.openUserProfile(), Reason.DomCallback);
-      return;
-    }
-    onAccountAction?.(action);
-  };
-
-  const handleSwitchSession = (sessionId: string) => {
-    detach(
-      clerk?.setActive({
-        session: sessionId,
-        beforeEmit: () => {
-          return window.location.reload();
-        },
-      }),
-      Reason.DomCallback,
-    );
-  };
-
-  const handleAddAccount = () => {
-    detach(clerk?.openSignIn(), Reason.DomCallback);
-  };
-
-  return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <button
-          type="button"
-          className={`rounded-lg transition-colors duration-200 ${
-            collapsed
-              ? "inline-flex h-8 w-8 shrink-0 items-center justify-center p-0 hover:bg-sidebar-accent/50"
-              : "flex w-full items-center gap-2 p-2 text-left hover:bg-sidebar-accent/50"
-          }`}
-        >
-          <AccountAvatar
-            imageUrl={user?.imageUrl}
-            name={accountName}
-            initial={accountInitial}
-          />
-          {!collapsed && (
-            <div className="flex-1 min-w-0">
-              <p className="text-sm font-medium leading-tight truncate text-sidebar-foreground">
-                {accountName}
-              </p>
-              <p className="text-xs leading-tight truncate mt-px text-sidebar-foreground opacity-70">
-                {accountEmail}
-              </p>
-            </div>
-          )}
-        </button>
-      </DropdownMenuTrigger>
-
-      <DropdownMenuContent
-        side="top"
-        align="start"
-        sideOffset={8}
-        className="w-[240px]"
-      >
-        {/* Current account header */}
-        {current && (
-          <>
-            <div className="px-3 py-3">
-              <div className="flex items-center gap-3">
-                <AccountAvatar
-                  imageUrl={current.imageUrl}
-                  name={current.name}
-                  initial={current.initial}
-                  size="md"
-                />
-                <div className="flex-1 min-w-0">
-                  <div className="text-sm font-medium text-foreground truncate">
-                    {current.name}
-                  </div>
-                  <div className="text-xs text-muted-foreground truncate">
-                    {current.email}
-                  </div>
-                </div>
-              </div>
-            </div>
-            <DropdownMenuSeparator />
-          </>
-        )}
-
-        {/* Preferences (standalone) */}
-        {!hidePreferences && (
-          <>
-            <DropdownMenuItem
-              onPointerDown={() => {
-                return handleAccountAction("preferences");
-              }}
-              className="gap-3 px-3 py-2.5 rounded-lg"
-            >
-              <IconAdjustmentsHorizontal
-                size={18}
-                stroke={1.5}
-                className="text-muted-foreground"
-              />
-              <span>Preferences</span>
-            </DropdownMenuItem>
-            <DropdownMenuSeparator />
-          </>
-        )}
-
-        {/* Account management group */}
-        {hasOthers ? (
-          <DropdownMenuSub>
-            <DropdownMenuSubTrigger className="gap-3 px-3 py-2.5 rounded-lg">
-              <IconSwitchHorizontal
-                size={18}
-                stroke={1.5}
-                className="text-muted-foreground"
-              />
-              <span className="flex-1">Switch account</span>
-              <IconChevronRight
-                size={14}
-                stroke={1.5}
-                className="text-muted-foreground"
-              />
-            </DropdownMenuSubTrigger>
-            <DropdownMenuSubContent className="w-[220px]">
-              {others.map((account) => {
-                return (
-                  <DropdownMenuItem
-                    key={account.sessionId}
-                    onPointerDown={() => {
-                      return handleSwitchSession(account.sessionId);
-                    }}
-                    className="gap-3 px-3 py-2.5 rounded-lg"
-                  >
-                    <AccountAvatar
-                      imageUrl={account.imageUrl}
-                      name={account.name}
-                      initial={account.initial}
-                    />
-                    <div className="flex-1 min-w-0">
-                      <div className="text-sm font-medium text-foreground truncate">
-                        {account.name}
-                      </div>
-                      <div className="text-xs text-muted-foreground truncate">
-                        {account.email}
-                      </div>
-                    </div>
-                  </DropdownMenuItem>
-                );
-              })}
-              <DropdownMenuSeparator />
-              <DropdownMenuItem
-                onPointerDown={handleAddAccount}
-                className="gap-3 px-3 py-2.5 rounded-lg"
-              >
-                <IconPlus
-                  size={18}
-                  stroke={1.5}
-                  className="text-muted-foreground"
-                />
-                <span>Add account</span>
-              </DropdownMenuItem>
-            </DropdownMenuSubContent>
-          </DropdownMenuSub>
-        ) : (
-          <DropdownMenuItem
-            onPointerDown={handleAddAccount}
-            className="gap-3 px-3 py-2.5 rounded-lg"
-          >
-            <IconPlus
-              size={18}
-              stroke={1.5}
-              className="text-muted-foreground"
-            />
-            <span>Add account</span>
-          </DropdownMenuItem>
-        )}
-        <DropdownMenuItem
-          onPointerDown={() => {
-            return handleAccountAction("manage");
-          }}
-          className="gap-3 px-3 py-2.5 rounded-lg"
-        >
-          <IconUser size={18} stroke={1.5} className="text-muted-foreground" />
-          <span>Manage account</span>
-        </DropdownMenuItem>
-        {showExportData && (
-          <DropdownMenuItem
-            onPointerDown={() => {
-              return window.open(`${apiBase}/export`, "_blank");
-            }}
-            className="gap-3 px-3 py-2.5 rounded-lg"
-          >
-            <IconDatabaseExport
-              size={18}
-              stroke={1.5}
-              className="text-muted-foreground"
-            />
-            <span>Export data</span>
-          </DropdownMenuItem>
-        )}
-        <DropdownMenuSeparator />
-        <DropdownMenuItem
-          onPointerDown={() => {
-            return handleAccountAction("signout");
-          }}
-          className="gap-3 px-3 py-2.5 rounded-lg"
-        >
-          <IconLogout
-            size={18}
-            stroke={1.5}
-            className="text-muted-foreground"
-          />
-          <span>Sign out</span>
-        </DropdownMenuItem>
-      </DropdownMenuContent>
-    </DropdownMenu>
-  );
-}
-
-function ChatThreadItem({
-  session,
-  isSelected,
-  onSelect,
-}: {
-  session: ChatThreadListItem;
-  isSelected: boolean;
-  onSelect?: (id: string) => void;
-}) {
-  const setPendingDeleteThreadId = useSet(setPendingDeleteThreadId$);
-
-  function handleDeleteClick(e: React.MouseEvent) {
-    e.preventDefault();
-    e.stopPropagation();
-    setPendingDeleteThreadId(session.id);
-  }
-
-  return (
-    <div className="group relative">
-      <Link
-        pathname="/chats/:id"
-        options={{ pathParams: { id: session.id } }}
-        onPointerDown={(e) => {
-          if (e.metaKey || e.ctrlKey || e.shiftKey) {
-            return;
-          }
-          e.preventDefault();
-          onSelect?.(session.id);
-        }}
-        className={`flex h-8 items-center gap-2 rounded-lg p-2 text-left text-sm leading-5 transition-colors ${
-          isSelected
-            ? "bg-gray-200 text-gray-900 font-medium"
-            : "text-sidebar-foreground hover:bg-sidebar-accent"
-        }`}
-      >
-        <span className="truncate min-w-0 flex-1">
-          {session.title ?? "New chat"}
-        </span>
-      </Link>
-      <div className="absolute right-0 top-0 flex h-8 w-8 items-center justify-center">
-        <TooltipProvider delayDuration={200}>
-          <Tooltip>
-            <TooltipTrigger asChild>
-              <button
-                type="button"
-                onPointerDown={handleDeleteClick}
-                className={`flex h-6 w-6 cursor-pointer items-center justify-center rounded-md invisible group-hover:visible transition-opacity duration-150 ${
-                  isSelected
-                    ? "text-slate-500 hover:text-slate-900 hover:bg-slate-300"
-                    : "text-sidebar-foreground/80 hover:text-foreground hover:bg-sidebar-foreground/10"
-                }`}
-                aria-label="Delete chat"
-              >
-                <IconTrash size={12} stroke={2} />
-              </button>
-            </TooltipTrigger>
-            <TooltipContent side="bottom">
-              <p className="text-xs">Delete chat</p>
-            </TooltipContent>
-          </Tooltip>
-        </TooltipProvider>
-      </div>
-    </div>
-  );
-}
-
-function RecentChatList({
-  loading,
-  error,
-  sessions,
-  searchTerm,
-  selectedRecentId,
-  onRecentSelect,
-}: {
-  loading: boolean;
-  error: string | null;
-  sessions: ChatThreadListItem[];
-  searchTerm: string;
-  selectedRecentId: string | null;
-  onRecentSelect?: (id: string) => void;
-}) {
-  const pendingDeleteThreadId = useGet(pendingDeleteThreadId$);
-  const setPendingDeleteThreadId = useSet(setPendingDeleteThreadId$);
-  const setDelete = useSet(deleteChatThread$);
-  const pageSignal = useGet(pageSignal$);
-
-  function confirmDelete() {
-    if (!pendingDeleteThreadId) {
-      return;
-    }
-    const threadId = pendingDeleteThreadId;
-    setPendingDeleteThreadId(null);
-    detach(setDelete(threadId, pageSignal), Reason.DomCallback);
-  }
-
-  if (loading && sessions.length === 0) {
-    return (
-      <>
-        {["w-3/4", "w-1/2", "w-2/3"].map((w) => {
-          return (
-            <div
-              key={w}
-              data-testid="sidebar-skeleton"
-              className="flex h-8 items-center rounded-lg p-2"
-            >
-              <Skeleton className={`h-4 ${w}`} />
-            </div>
-          );
-        })}
-      </>
-    );
-  }
-  if (error) {
-    return (
-      <p
-        data-testid="chat-threads-error"
-        className="px-2 py-2 text-xs text-destructive"
-      >
-        {error}
-      </p>
-    );
-  }
-  if (sessions.length === 0) {
-    return (
-      <p className="px-2 py-2 text-xs text-muted-foreground/70 leading-relaxed">
-        {searchTerm.trim()
-          ? "No chats match your search"
-          : "Start a conversation and it'll show up here"}
-      </p>
-    );
-  }
-  return (
-    <>
-      {sessions.map((session) => {
-        return (
-          <ChatThreadItem
-            key={session.id}
-            session={session}
-            isSelected={selectedRecentId === session.id}
-            onSelect={onRecentSelect}
-          />
-        );
-      })}
-      <Dialog
-        open={pendingDeleteThreadId !== null}
-        onOpenChange={(open) => {
-          if (!open) {
-            setPendingDeleteThreadId(null);
-          }
-        }}
-      >
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Delete chat?</DialogTitle>
-            <DialogDescription>
-              This will permanently delete this chat. This action cannot be
-              undone.
-            </DialogDescription>
-          </DialogHeader>
-          <DialogFooter>
-            <Button
-              variant="outline"
-              onPointerDown={() => {
-                setPendingDeleteThreadId(null);
-              }}
-            >
-              Cancel
-            </Button>
-            <Button variant="destructive" onPointerDown={confirmDelete}>
-              Delete
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
-    </>
-  );
-}
-
-function RecentChatSection({
-  currentChatAgentId,
-  displayName,
-  subagents,
-  selectedRecentId,
-  onRecentSelect,
-  onNewChat,
-  newChatDisabled,
-}: {
-  currentChatAgentId: string | null;
-  displayName: string;
-  subagents: SubagentInfo[];
-  selectedRecentId: string | null;
-  onRecentSelect?: (id: string) => void;
-  onNewChat?: (agentId: string | null) => void;
-  newChatDisabled?: boolean;
-}) {
-  const recentSessionsLoadable = useLastLoadable(chatThreads$);
-  const recentSessions =
-    recentSessionsLoadable.state === "hasData"
-      ? recentSessionsLoadable.data
-      : [];
-  const recentSessionsLoading = recentSessionsLoadable.state === "loading";
-  const recentSessionsError =
-    recentSessionsLoadable.state === "hasError"
-      ? recentSessionsLoadable.error instanceof Error
-        ? recentSessionsLoadable.error.message
-        : "Failed to load chats"
-      : null;
-  const searchOpen = useGet(sidebarSearchOpen$);
-  const setSearchOpen = useSet(setSidebarSearchOpen$);
-  const searchTerm = useGet(sidebarSearchTerm$);
-  const setSearchTerm = useSet(setSidebarSearchTerm$);
-  const collapsed = useGet(sessionListCollapsed$);
-  const setCollapsed = useSet(setSessionListCollapsed$);
-
-  // Filter sessions by current agent
-  const subagentIds = new Set(
-    subagents.map((a) => {
-      return a.id;
-    }),
-  );
-  const agentSessions = currentChatAgentId
-    ? recentSessions.filter((s) => {
-        return s.agentId === currentChatAgentId;
-      })
-    : recentSessions.filter((s) => {
-        return !subagentIds.has(s.agentId);
-      });
-
-  const matchedAgent = subagents.find((a) => {
-    return a.id === currentChatAgentId;
-  });
-  const agentLabel = currentChatAgentId
-    ? (matchedAgent?.displayName ?? matchedAgent?.id ?? displayName)
-    : displayName;
-
-  const trimmedTerm = searchTerm.trim().toLowerCase();
-  const filteredSessions = trimmedTerm
-    ? agentSessions.filter((s) => {
-        return (s.title ?? "").toLowerCase().includes(trimmedTerm);
-      })
-    : agentSessions;
-
-  const handleNewChat = onNewChat
-    ? () => {
-        onNewChat(currentChatAgentId ?? null);
-      }
-    : undefined;
-
-  return (
-    <div className="mt-4 flex flex-col">
-      {searchOpen ? (
-        <div
-          className="shrink-0 flex h-8 items-center gap-2 rounded-lg bg-sidebar-accent/60 pl-2 pr-2 zero-border"
-          onBlur={(e) => {
-            if (!e.currentTarget.contains(e.relatedTarget)) {
-              setSearchOpen(false);
-            }
-          }}
-        >
-          <IconSearch
-            size={15}
-            stroke={2.5}
-            className="shrink-0 text-sidebar-foreground/50"
-          />
-          <input
-            type="text"
-            value={searchTerm}
-            onChange={(e) => {
-              return setSearchTerm(e.target.value);
-            }}
-            placeholder={`Search chat with ${agentLabel}`}
-            autoFocus
-            className="flex-1 min-w-0 bg-transparent text-sm leading-5 text-sidebar-foreground placeholder:text-sidebar-foreground/50 focus:outline-none"
-          />
-          <div className="flex h-8 w-8 shrink-0 items-center justify-center">
-            <button
-              type="button"
-              onPointerDown={() => {
-                setSearchOpen(false);
-              }}
-              className="shrink-0 flex items-center justify-center h-5 w-5 rounded text-sidebar-foreground/50 hover:text-sidebar-foreground hover:bg-sidebar-accent transition-colors"
-              aria-label="Close search"
-            >
-              <IconX size={12} stroke={2} />
-            </button>
-          </div>
-        </div>
-      ) : (
-        <div
-          className="zero-nav-recent-label group flex h-8 shrink-0 cursor-pointer items-center justify-between rounded-lg pl-2 pr-0 hover:bg-sidebar-accent/50 transition-colors"
-          onPointerDown={() => {
-            return setCollapsed(!collapsed);
-          }}
-        >
-          <span className="flex flex-1 items-center gap-1 truncate text-[13px] font-medium leading-4 text-sidebar-foreground/50 group-hover:text-sidebar-foreground transition-colors">
-            Chats with {agentLabel}
-            <span className="shrink-0 opacity-0 group-hover:opacity-100 transition-opacity duration-200">
-              <IconChevronRight
-                size={12}
-                stroke={2}
-                className={collapsed ? "" : "rotate-90"}
-              />
-            </span>
-          </span>
-          <div className="flex items-center gap-0.5">
-            <TooltipProvider delayDuration={200}>
-              <Tooltip>
-                <TooltipTrigger asChild>
-                  <button
-                    type="button"
-                    onPointerDown={(e) => {
-                      e.stopPropagation();
-                      setSearchOpen(true);
-                    }}
-                    className="relative z-10 flex h-8 w-8 items-center justify-center rounded-lg text-sidebar-foreground/70 hover:text-sidebar-foreground hover:bg-sidebar-accent transition-colors"
-                    aria-label="Search chats"
-                  >
-                    <IconSearch size={15} stroke={2.5} />
-                  </button>
-                </TooltipTrigger>
-                <TooltipContent side="bottom">
-                  <p className="text-xs">Search chats</p>
-                </TooltipContent>
-              </Tooltip>
-            </TooltipProvider>
-            {handleNewChat && (
-              <TooltipProvider delayDuration={200}>
-                <Tooltip>
-                  <TooltipTrigger asChild>
-                    <button
-                      type="button"
-                      onPointerDown={(e) => {
-                        e.stopPropagation();
-                        handleNewChat();
-                      }}
-                      disabled={newChatDisabled}
-                      className="relative z-10 flex h-8 w-8 items-center justify-center rounded-lg text-sidebar-foreground/70 hover:text-sidebar-foreground hover:bg-sidebar-accent transition-colors disabled:opacity-50 disabled:pointer-events-none"
-                      aria-label={`New chat with ${agentLabel}`}
-                    >
-                      <IconPlus size={15} stroke={2.5} />
-                    </button>
-                  </TooltipTrigger>
-                  <TooltipContent side="bottom">
-                    <p className="text-xs">New chat</p>
-                  </TooltipContent>
-                </Tooltip>
-              </TooltipProvider>
-            )}
-          </div>
-        </div>
-      )}
-      {!collapsed && (
-        <div className="mt-1">
-          <div className="flex flex-col gap-1">
-            <RecentChatList
-              loading={recentSessionsLoading}
-              error={recentSessionsError}
-              sessions={filteredSessions}
-              searchTerm={searchTerm}
-              selectedRecentId={selectedRecentId}
-              onRecentSelect={onRecentSelect}
-            />
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function TalkToSection({
-  activeId,
-  currentChatAgentId,
-  selectedRecentId,
-  selectedAgentIdFromChat,
-  displayName,
-  defaultAgentRawName,
-  zeroAvatarSrc,
-  pinnedAgents,
-  pinnedIds,
-  subagents,
-  onPinnedIdsChange,
-  onNewChat,
-}: {
-  activeId: RouteKey | null;
-  currentChatAgentId: string | null;
-  selectedRecentId: string | null;
-  selectedAgentIdFromChat: string | null | undefined;
-  displayName: string;
-  defaultAgentRawName?: string | null;
-  zeroAvatarSrc: string | null;
-  pinnedAgents: SubagentInfo[];
-  pinnedIds: string[];
-  subagents: SubagentInfo[];
-  onPinnedIdsChange: (ids: string[]) => void;
-  onNewChat?: (agentId: string | null) => void;
-}) {
-  const chatListOpen = useGet(chatListOpen$);
-  const setChatListOpenFn = useSet(setChatListOpen$);
-  const collapsed = useGet(agentCardCollapsed$);
-  const setCollapsed = useSet(setAgentCardCollapsed$);
-  const reloadAgents = useSet(reloadAgents$);
-
-  return (
-    <div className="shrink-0">
-      <div
-        className="group flex h-8 cursor-pointer items-center justify-between rounded-lg pl-2 pr-0 hover:bg-sidebar-accent/50 transition-colors"
-        data-testid="pinned-section-header"
-        onPointerDown={() => {
-          return setCollapsed(!collapsed);
-        }}
-      >
-        <span className="flex flex-1 items-center gap-1 truncate text-[13px] font-medium leading-4 text-sidebar-foreground/50 group-hover:text-sidebar-foreground transition-colors">
-          Pinned
-          <span className="shrink-0 opacity-0 group-hover:opacity-100 transition-opacity duration-200">
-            <IconChevronRight
-              size={12}
-              stroke={2}
-              className={collapsed ? "" : "rotate-90"}
-            />
-          </span>
-        </span>
-        <TooltipProvider delayDuration={200}>
-          <Tooltip>
-            <TooltipTrigger asChild>
-              <button
-                type="button"
-                onPointerDown={(e) => {
-                  e.stopPropagation();
-                  setChatListOpenFn(true);
-                  reloadAgents();
-                }}
-                className="relative z-10 flex h-8 w-8 items-center justify-center rounded-lg text-sidebar-foreground/70 hover:text-sidebar-foreground hover:bg-sidebar-accent transition-colors"
-                aria-label="Open a conversation"
-              >
-                <IconPlus size={15} stroke={2.5} />
-              </button>
-            </TooltipTrigger>
-            <TooltipContent side="right">
-              <p className="text-xs">Open a conversation</p>
-            </TooltipContent>
-          </Tooltip>
-        </TooltipProvider>
-      </div>
-      {!collapsed && (
-        <div className="flex flex-col gap-0.5 mt-1">
-          {/* Lead agent */}
-          {(() => {
-            const isPrimarySelected =
-              isChatRoute(activeId) &&
-              !selectedRecentId &&
-              currentChatAgentId === null;
-            const isFromChat =
-              selectedAgentIdFromChat !== undefined &&
-              selectedAgentIdFromChat === null;
-            return (
-              <Link
-                pathname={defaultAgentRawName ? "/agents/:id/chat" : "/"}
-                options={
-                  defaultAgentRawName
-                    ? { pathParams: { id: defaultAgentRawName } }
-                    : undefined
-                }
-                className={`flex w-full h-8 shrink-0 items-center gap-2 rounded-lg px-2 text-left text-sm leading-5 no-underline transition-colors duration-200 ${
-                  isPrimarySelected
-                    ? "bg-gray-200 text-gray-900 font-medium"
-                    : isFromChat
-                      ? "border-l-2 border-[hsl(var(--gray-400))] bg-sidebar-accent/50"
-                      : "text-sidebar-foreground hover:bg-sidebar-accent"
-                }`}
-              >
-                {zeroAvatarSrc ? (
-                  <img
-                    src={zeroAvatarSrc}
-                    alt={displayName}
-                    className="h-5 w-5 shrink-0 rounded-md object-cover object-top"
-                  />
-                ) : (
-                  <div
-                    className="h-5 w-5 shrink-0 rounded-md bg-muted"
-                    aria-hidden
-                  />
-                )}
-                <span className="truncate">{displayName}</span>
-              </Link>
-            );
-          })()}
-          {/* Pinned agents */}
-          {pinnedAgents.map((agent) => {
-            const isPrimarySelected =
-              isChatRoute(activeId) &&
-              !selectedRecentId &&
-              currentChatAgentId === agent.id;
-            const isFromChat = selectedAgentIdFromChat === agent.id;
-            return (
-              <div
-                key={agent.id}
-                className="group relative"
-                data-testid="pinned-agent-card"
-              >
-                <Link
-                  pathname="/agents/:id/chat"
-                  options={{ pathParams: { id: agent.id } }}
-                  className={`flex w-full h-8 shrink-0 items-center gap-2 rounded-lg px-2 text-left text-sm leading-5 no-underline transition-colors duration-200 ${
-                    isPrimarySelected
-                      ? "bg-gray-200 text-gray-900 font-medium"
-                      : isFromChat
-                        ? "border-l-2 border-[hsl(var(--gray-400))] bg-sidebar-accent/50"
-                        : "text-sidebar-foreground hover:bg-sidebar-accent"
-                  }`}
-                >
-                  <AgentAvatarImg
-                    name={agent.id}
-                    alt={agent.displayName ?? agent.id}
-                    className="h-5 w-5 shrink-0 rounded-md object-cover object-top"
-                  />
-                  <span className="truncate">
-                    {agent.displayName ?? agent.id}
-                  </span>
-                </Link>
-                <div className="absolute right-0 top-0 flex h-8 w-8 items-center justify-center">
-                  <TooltipProvider delayDuration={200}>
-                    <Tooltip>
-                      <TooltipTrigger asChild>
-                        <button
-                          type="button"
-                          onPointerDown={(e) => {
-                            e.preventDefault();
-                            e.stopPropagation();
-                            onPinnedIdsChange(
-                              pinnedIds.filter((id) => {
-                                return id !== agent.id;
-                              }),
-                            );
-                          }}
-                          className={`flex h-6 w-6 cursor-pointer items-center justify-center rounded-md invisible group-hover:visible transition-opacity duration-150 ${
-                            isPrimarySelected
-                              ? "text-slate-500 hover:text-slate-900 hover:bg-slate-300"
-                              : "text-sidebar-foreground/80 hover:text-foreground hover:bg-sidebar-foreground/10"
-                          }`}
-                          aria-label="Remove from list"
-                        >
-                          <IconX size={12} stroke={2} />
-                        </button>
-                      </TooltipTrigger>
-                      <TooltipContent side="right">
-                        <p className="text-xs">Remove from list</p>
-                      </TooltipContent>
-                    </Tooltip>
-                  </TooltipProvider>
-                </div>
-              </div>
-            );
-          })}
-        </div>
-      )}
-
-      <ChatListDialog
-        open={chatListOpen}
-        onOpenChange={setChatListOpenFn}
-        zeroAvatarSrc={zeroAvatarSrc}
-        displayName={displayName}
-        subagents={subagents}
-        pinnedIds={pinnedIds}
-        onPinnedIdsChange={onPinnedIdsChange}
-        onNewChat={onNewChat}
-      />
-    </div>
-  );
-}
-
-/** Overlay scroll area: hides native scrollbar, renders a custom thin indicator. */
-function OverlayScrollArea({
-  className,
-  children,
-  onScroll,
-  style,
-  "data-testid": dataTestId,
-}: {
-  className?: string;
-  children: ReactNode;
-  onScroll?: (e: React.UIEvent<HTMLDivElement>) => void;
-  style?: React.CSSProperties;
-  "data-testid"?: string;
-}) {
-  const thumbStyleValue = useGet(thumbStyle$);
-  const setThumbStyleFn = useSet(setThumbStyle$);
-  const hovering = useGet(hovering$);
-  const setHoveringFn = useSet(setHovering$);
-
-  const handleScroll = (e: React.UIEvent<HTMLDivElement>) => {
-    onScroll?.(e);
-    const el = e.currentTarget;
-    const { scrollTop, scrollHeight, clientHeight } = el;
-    if (scrollHeight <= clientHeight) {
-      setThumbStyleFn({
-        top: thumbStyleValue.top,
-        height: thumbStyleValue.height,
-        visible: false,
-      });
-      return;
-    }
-    const ratio = clientHeight / scrollHeight;
-    const thumbH = Math.max(ratio * clientHeight, 24);
-    const maxTop = clientHeight - thumbH;
-    const top = (scrollTop / (scrollHeight - clientHeight)) * maxTop;
-    setThumbStyleFn({ top, height: thumbH, visible: true });
-  };
-
-  const showThumb = thumbStyleValue.visible && hovering;
-
-  return (
-    <div
-      className={`relative ${className ?? ""}`}
-      onMouseEnter={() => {
-        return setHoveringFn(true);
-      }}
-      onMouseLeave={() => {
-        return setHoveringFn(false);
-      }}
-    >
-      <div
-        className="h-full overflow-y-auto overflow-x-hidden [scrollbar-width:none] [&::-webkit-scrollbar]:hidden"
-        style={style}
-        onScroll={handleScroll}
-        data-testid={dataTestId}
-      >
-        {children}
-      </div>
-      <div
-        className="absolute -right-2 top-0 bottom-0 w-[6px] pointer-events-none"
-        aria-hidden="true"
-        style={{ opacity: showThumb ? 1 : 0, transition: "opacity 150ms" }}
-      >
-        <div
-          className="absolute right-0 w-[5px] rounded-full bg-foreground/15"
-          style={{ top: thumbStyleValue.top, height: thumbStyleValue.height }}
-        />
-      </div>
-    </div>
-  );
-}
-
-function nextTierInfo(tier: string): { label: string; img: string } | null {
-  if (tier === "free") {
-    return { label: "Pro", img: planProImg };
-  }
-  if (tier === "pro") {
-    return { label: "Team", img: planTeamImg };
-  }
-  return null;
-}
-
-function SidebarUpgradeCard() {
-  const pageSignal = useGet(pageSignal$);
-  const billingLoadable = useLastLoadable(billingStatusAsync$);
-  const billing =
-    billingLoadable.state === "hasData" ? billingLoadable.data : null;
-  const isAdminLoadable = useLoadable(isOrgAdmin$);
-  const isAdmin =
-    isAdminLoadable.state === "hasData" ? isAdminLoadable.data : false;
-  const setTab = useSet(setActiveTab$);
-  const setSubPage = useSet(setBillingSubPage$);
-  const openManage = useSet(setOrgManageDialogOpen$);
-
-  if (!isAdmin) {
-    return null;
-  }
-
-  if (!billing) {
-    return null;
-  }
-  const next = nextTierInfo(billing.tier);
-  if (!next) {
-    return null;
-  }
-
-  const handleClick = () => {
-    setTab("billing");
-    setSubPage(true);
-    detach(openManage(true, pageSignal), Reason.DomCallback);
-  };
-
-  return (
-    <button
-      type="button"
-      onPointerDown={handleClick}
-      className="flex w-full items-center gap-3 rounded-lg p-2.5 text-left transition-colors hover:bg-muted/30 zero-card shadow-[0_1px_2px_hsl(220_12%_20%/0.04),0_4px_12px_hsl(220_12%_20%/0.03)]"
-    >
-      <div className="min-w-0 flex-1">
-        <p className="text-sm font-medium text-foreground">Get {next.label}</p>
-        <p className="text-[11px] text-muted-foreground mt-0.5">
-          More credits & active agents
-        </p>
-      </div>
-      <img
-        src={next.img}
-        alt={next.label}
-        className="h-14 w-14 shrink-0 object-contain -my-3"
-      />
-    </button>
-  );
-}
 
 export function ZeroSidebar() {
   const isScrolled = useGet(isScrolled$);
